meson: Check for debug and optimization in the same way as GTK 4
Now meson returns 'custom' as the 'buildtype' when the user sets
'optimization' and 'debug' to a combination it does not know, making it
more likely to have none of 'debug' and 'release' in 'buildtype'.
Therefore, it should be better to use 'optimization' and 'debug' options
directly if we don't want an explicit option to control it.